services/tujur/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "tujur"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "pydantic>=2.7",
    "anthropic>=0.28",
    "psycopg[binary,pool]>=3.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["app*", "adapters*", "domain*"]
namespaces = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
